+++ application/core/CSV_Model.php
@@ -87,8 +87,9 @@ class CSV_Model extends Memory_Model
 		if (($handle = fopen($this->_origin, "w")) !== FALSE)
 		{
 			fputcsv($handle, $this->_fields);
-			foreach ($this->_data as $key => $record)
+			foreach ($this->_data as $key => $record) {
 				fputcsv($handle, array_values((array) $record));
+			}
 			fclose($handle);
 		}
 		// --------------------